Narrow the scope of the write lock in the remove function.
It seems that in some cases, calling item.property can result in a call
back into this model's index() function. This obviously results in a
deadlock since we hold the write lock at this point. The fix is to move
the write lock down a bit so it doesn't block the call to property().