Add a function 'reSort()' to 'Collection',
which can be called from QML, as well.

The function 'reSort()' triggers sorting the collection again.
Calling it is *not* required if data within the model itself has changed
(re-sorting will be conducted automatically in that case).
You want to call this function if external data has been updated,
e.g. you sort a model by distance and the current position
of the user has changed.